requirements.txt
gradio
nltk
numpy
pandas
Pillow
pycocoevalcap
pytorch_lightning
rouge_score
spacy
torch
torchtext
torchvision
tensorboardX